backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30
sqlalchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.0
python-jose>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0
pydantic[email]
